Make Spring Data Rest optional again
