+++ spring-content-rest/src/main/java/internal/org/springframework/content/rest/controllers/ResourceHandlerMethodArgumentResolver.java
@@ -7,12 +7,11 @@ import org.springframework.content.commons.repository.Store;
 import org.springframework.content.commons.storeservice.StoreInfo;
 import org.springframework.content.commons.storeservice.Stores;
 import org.springframework.content.rest.config.RestConfiguration;
+import org.springframework.context.ApplicationContext;
 import org.springframework.core.MethodParameter;
 import org.springframework.core.io.Resource;
 import org.springframework.data.repository.support.Repositories;
 import org.springframework.data.repository.support.RepositoryInvokerFactory;
-import org.springframework.data.rest.webmvc.BaseUri;
-import org.springframework.data.rest.webmvc.config.RootResourceInformationHandlerMethodArgumentResolver;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.ModelAndViewContainer;
@@ -25,8 +24,8 @@ import internal.org.springframework.content.rest.utils.StoreUtils;
 
 public class ResourceHandlerMethodArgumentResolver extends StoreHandlerMethodArgumentResolver {
 
-    public ResourceHandlerMethodArgumentResolver(RestConfiguration config, Repositories repositories, RepositoryInvokerFactory repoInvokerFactory, Stores stores, RootResourceInformationHandlerMethodArgumentResolver rootResourceInfoResolver, BaseUri baseUri) {
-        super(config, repositories, repoInvokerFactory, stores, rootResourceInfoResolver, baseUri);
+    public ResourceHandlerMethodArgumentResolver(ApplicationContext context, RestConfiguration config, Repositories repositories, RepositoryInvokerFactory repoInvokerFactory, Stores stores) {
+        super(context, config, repositories, repoInvokerFactory, stores);
     }
 
     @Override
