+++ spring-content-rest/src/main/java/org/springframework/content/rest/config/RestConfiguration.java
@@ -20,8 +20,6 @@ import org.springframework.core.io.Resource;
 import org.springframework.data.repository.support.DefaultRepositoryInvokerFactory;
 import org.springframework.data.repository.support.Repositories;
 import org.springframework.data.repository.support.RepositoryInvokerFactory;
-import org.springframework.data.rest.webmvc.BaseUri;
-import org.springframework.data.rest.webmvc.config.RootResourceInformationHandlerMethodArgumentResolver;
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
@@ -131,11 +129,11 @@ public class RestConfiguration implements InitializingBean {
 		@Autowired
 		private StoreByteRangeHttpRequestHandler byteRangeRestRequestHandler;
 
-		@Autowired
-		private RootResourceInformationHandlerMethodArgumentResolver rootResourceInfoResolver;
+//		@Autowired
+//		private RootResourceInformationHandlerMethodArgumentResolver rootResourceInfoResolver;
 
-		@Autowired
-		private BaseUri baseUri;
+//		@Autowired
+//		private BaseUri baseUri;
 
 		@Autowired
 		private Stores stores;
@@ -143,7 +141,7 @@ public class RestConfiguration implements InitializingBean {
 		@Override
 		public void addArgumentResolvers(List<HandlerMethodArgumentResolver> argumentResolvers) {
 
-			argumentResolvers.add(new ResourceHandlerMethodArgumentResolver(config, repositories, repoInvokerFactory, stores, rootResourceInfoResolver, baseUri));
+			argumentResolvers.add(new ResourceHandlerMethodArgumentResolver(context, config, repositories, repoInvokerFactory, stores));
 		}
 
 		@Override
